+++ packages/docs/src/content/regles-validation/min-length.md
@@ -1,46 +1,40 @@
 ---
 title: Règle de validation minLength
-description : La règle `minLength` permet d’imposer la longueur minimun d’une chaîne de caractère.
+description: La règle `minLength` permet d’imposer la longueur minimale d’un champ de formulaire.
 ---
 
-## Utilisation
+<doc-tabs>
 
-<doc-indent>
+<doc-tab-item label="Utilisation">
 
 Pour utiliser la règle de validation `minLength`, vous devez l'importer :
-
-</doc-indent>
 
 ```ts
 import { minLength } from '@cnamts/vue-dot/src/rules/minLength';
 ```
 
-Puis vous devez l’appeler avec différents arguments :
-
-```ts
-minLengthRule = minLength(min, errorMessages)
-```
-
-<doc-api name="rules/min-length/arguments"></doc-api>
-
-Puis l'ajouter dans un tableau qui sera utilisé sur un champ de formulaire via la prop `rules` :
+Puis l'ajouter dans un tableau qui sera utilisé sur un champ de formulaire via la prop `rules` et passer en argument la longueur minimale que le champ doit avoir :
 
 ```ts
 rules = [
-    minLengthRule
+	minLength(lengthValue)
 ];
 ```
 
-## Modifier les messages d’erreur
+## Messages d’erreur
 
-<doc-indent>
-
-Il est possible de modifier les messages d’erreur en utilisant l’argument `errorMessages`. Il s’agit d’un objet qui contient le message d’erreur utilisé par cette règle de validation :
-
-</doc-indent>
+Vous pouvez modifier les messages d’erreur par défaut en passant en argument un objet contenant les messages d’erreur utilisés par cette règle de validation :
 
 ```ts
-minLengthRule = minLength(lengthValue, ignoreSpaces, {
-    default: (min: number) => `${min} caractères minimum`
+minLengthRule = minLength(lengthValue, {
+	default: (min: number) => `La longueur minimale du champ est de ${min} caractères.`
 });
 ```
+
+</doc-tab-item>
+
+<doc-tab-item label="API">
+<doc-api name="rules/min-length"></doc-api>
+</doc-tab-item>
+
+</doc-tabs>
