--- packages/docs/src/data/api/rules/min-length/arguments.ts
@@ -1,18 +1,0 @@
-import { Api } from '~/types';
-
-export const api: Api = {
-	exactLength: {
-		arguments: [
-			{
-				name: 'min',
-				type: 'number',
-				description: 'La longueur minimum souhaitée'
-			},
-			{
-				name: 'errorMessages',
-				type: 'object',
-				description: 'Message d’erreurs'
-			}
-		]
-	}
-};
